Refactor Pomodoro class, streamline program execution
This commit populates the Pomodoro class instead of having most of the
code live in the executable pom-poem.rb file. This allows a user to make
more than one pom in an easy way. Likewise, a loop was added at the end
of the program to keep creating pom's if the user wants to and only
exits if the user specifically chooses to.